fixes view for show advertisement

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,7 +15,7 @@ class CreateUsersTable extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->tinyInteger('number');
+            $table->bigInteger('number');
             $table->string('email')->unique();
             $table->boolean('isadmin');
             $table->string('password');
